move code out of Collection Class
